Update method return type

+++ client/src/main/java/io/enjincoin/sdk/client/service/identity/vo/TokenEntry.java
@@ -24,7 +24,7 @@ public class TokenEntry {
     /**
      * @return the token id
      */
-    public long getTokenId() {
+    public int getTokenId() {
         return this.tokenId;
     }
